refactor(cli): share the framed-protocol test helpers between bridge E2Es
Copilot: the SYNC|LEN|payload|CRC16 mirror (crc16/build/parse) was
duplicated in serve_stream_bridge_e2e and serve_kble_e2e; extract it to
tests/framed_protocol/mod.rs so future wire-format tweaks stay in one
place.
